deleted all references in Selenium that are now in utilities reconnected all references
