+++ selenium/pom.xml
@@ -217,5 +217,11 @@
             <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.deque.html.axe-core</groupId>
+            <artifactId>dequeutilites</artifactId>
+            <version>4.4.1</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
