--- selenium/src/main/java/com/deque/html/axecore/results/Node.java
@@ -1,34 +1,0 @@
-package com.deque.html.axecore.results;
-
-public class Node {
-  private String html;
-  private Object target;
-
-  public String getHtml() {
-    return html;
-  }
-
-  public void setHtml(final String html) {
-    this.html = html;
-  }
-
-  public Object getTarget() {
-    return target;
-  }
-
-  public void setTarget(final Object target) {
-    this.target = target;
-  }
-
-  @Override
-  public String toString() {
-    return "Node{"
-        + "target='"
-        + target
-        + '\''
-        + ", html='"
-        + html
-        + '\''
-        + '}';
-  }
-}
