--- selenium/src/main/java/com/deque/html/axecore/results/TestEngine.java
@@ -1,18 +1,0 @@
-package com.deque.html.axecore.results;
-
-public class TestEngine {
-  private String name;
-  private String version;
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getVersion() {
-    return version;
-  }
-}
