--- selenium/src/main/java/com/deque/html/axecore/results/TestEnvironment.java
@@ -1,29 +1,0 @@
-package com.deque.html.axecore.results;
-
-public class TestEnvironment {
-  private String userAgent;
-  private int windowWidth;
-  private int windowHeight;
-  private double orientationAngle;
-  private String orientationType;
-
-  public String getUserAgent() {
-    return userAgent;
-  }
-
-  public int getWindowHeight() {
-    return windowHeight;
-  }
-
-  public int getwindowWidth() {
-    return windowWidth;
-  }
-
-  public double getOrientationAngle() {
-    return orientationAngle;
-  }
-
-  public String getOrientationType() {
-    return orientationType;
-  }
-}
