--- selenium/src/main/java/com/deque/html/axecore/results/TestRunner.java
@@ -1,9 +1,0 @@
-package com.deque.html.axecore.results;
-
-public class TestRunner {
-  private String name;
-
-  public String getName() {
-    return name;
-  }
-}
