+++ selenium/src/test/java/com/deque/html/axecore/selenium/AxeBuilderUnitTest.java
@@ -12,12 +12,6 @@
 
 package com.deque.html.axecore.selenium;
 
-import com.deque.html.axecore.axeargs.AxeRuleOptions;
-import com.deque.html.axecore.axeargs.AxeRunContext;
-import com.deque.html.axecore.axeargs.AxeRunOnlyOptions;
-import com.deque.html.axecore.axeargs.AxeRunOptions;
-import com.deque.html.axecore.results.Results;
-import com.deque.html.axecore.results.Rule;
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
@@ -27,6 +21,13 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import javax.naming.OperationNotSupportedException;
+
+import com.deque.html.axecore.args.AxeRuleOptions;
+import com.deque.html.axecore.args.AxeRunContext;
+import com.deque.html.axecore.args.AxeRunOnlyOptions;
+import com.deque.html.axecore.args.AxeRunOptions;
+import com.deque.html.axecore.results.Results;
+import com.deque.html.axecore.results.Rule;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
