+++ selenium/src/test/java/com/deque/html/axecore/selenium/AxeIntegrationUnitTest.java
@@ -12,9 +12,6 @@
 
 package com.deque.html.axecore.selenium;
 
-import com.deque.html.axecore.axeargs.AxeRunOptions;
-import com.deque.html.axecore.results.Results;
-import com.deque.html.axecore.results.Rule;
 import java.io.File;
 import java.io.IOException;
 import java.time.Duration;
@@ -25,6 +22,10 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import javax.naming.OperationNotSupportedException;
+
+import com.deque.html.axecore.args.AxeRunOptions;
+import com.deque.html.axecore.results.Results;
+import com.deque.html.axecore.results.Rule;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
